requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
Werkzeug==3.0.3
pandas>=2.2.0
openpyxl==3.1.5
python-dotenv==1.0.1
email-validator==2.1.1
requests==2.31.0
aiohttp==3.9.5
plotly>=5.0.0
python-docx==0.8.11
bleach==6.0.0
reportlab==4.0.5
